[Refactor] Align CSS parser numeric types
- Use size_t throughout tokenizer offset and range operations to match string sizes.
- Make clamped selector coefficient conversions explicit and avoid narrowing diagnostics.

issue: f-7350207575
AutoSubmit: true

+++ core/renderer/css/ng/parser/css_tokenizer.cc
@@ -302,7 +302,7 @@ CSSParserToken CSSTokenizer::ConsumeNumber() {
 
   NumericValueType type = kIntegerValueType;
   NumericSign sign = kNoSign;
-  unsigned number_length = 0;
+  size_t number_length = 0;
 
   UChar next = input_.PeekWithoutReplacement(0);
   if (next == '+') {
@@ -373,10 +373,10 @@ CSSParserToken CSSTokenizer::ConsumeIdentLikeToken() {
 // https://drafts.csswg.org/css-syntax/#consume-a-string-token
 CSSParserToken CSSTokenizer::ConsumeStringTokenUntil(UChar ending_code_point) {
   // Strings without escapes get handled without allocations
-  for (unsigned size = 0;; size++) {
+  for (size_t size = 0;; size++) {
     UChar cc = input_.PeekWithoutReplacement(size);
     if (cc == ending_code_point) {
-      unsigned start_offset = input_.Offset();
+      size_t start_offset = input_.Offset();
       input_.Advance(size + 1);
       return CSSParserToken(kStringToken, input_.RangeAt(start_offset, size));
     }
@@ -454,10 +454,10 @@ CSSParserToken CSSTokenizer::ConsumeUrlToken() {
   input_.AdvanceUntilNonWhitespace();
 
   // URL tokens without escapes get handled without allocations
-  for (unsigned size = 0;; size++) {
+  for (size_t size = 0;; size++) {
     UChar cc = input_.PeekWithoutReplacement(size);
     if (cc == ')') {
-      unsigned start_offset = input_.Offset();
+      size_t start_offset = input_.Offset();
       input_.Advance(size + 1);
       return CSSParserToken(kUrlToken, input_.RangeAt(start_offset, size));
     }
@@ -540,7 +540,7 @@ bool CSSTokenizer::ConsumeIfNext(UChar character) {
 // http://www.w3.org/TR/css3-syntax/#consume-a-name
 std::u16string CSSTokenizer::ConsumeName() {
   // Names without escapes get handled without allocations
-  for (unsigned size = 0;; ++size) {
+  for (size_t size = 0;; ++size) {
     UChar cc = input_.PeekWithoutReplacement(size);
     if (IsNameCodePoint(cc)) continue;
     // peekWithoutReplacement will return NUL when we hit the end of the
@@ -548,7 +548,7 @@ std::u16string CSSTokenizer::ConsumeName() {
     // below.
     if (cc == '\0' && input_.Offset() + size < input_.length()) break;
     if (cc == '\\') break;
-    unsigned start_offset = input_.Offset();
+    size_t start_offset = input_.Offset();
     input_.Advance(size);
     return input_.RangeAt(start_offset, size);
   }
